code _cleaning
Code tested and cleaned. New folders with specific libraries added
