+++ src/communication/spi.c
@@ -1,0 +1,99 @@
+/*
+ * spi.c
+ *
+ *  Created on: 18. 12. 2016
+ *      Author: Miroslav Kohút
+ */
+#include<communication/spi.h>
+
+void init_SPI1(void){
+
+	GPIO_InitTypeDef GPIO_InitStruct;
+	SPI_InitTypeDef SPI_InitStruct;
+
+	RCC_AHBPeriphClockCmd(RCC_AHBPeriph_GPIOA, ENABLE);
+
+	GPIO_InitStruct.GPIO_Pin = GPIO_Pin_7 | GPIO_Pin_6 | GPIO_Pin_5;
+	GPIO_InitStruct.GPIO_Mode = GPIO_Mode_AF;
+	GPIO_InitStruct.GPIO_OType = GPIO_OType_PP;
+	GPIO_InitStruct.GPIO_Speed = GPIO_Speed_40MHz;
+	GPIO_InitStruct.GPIO_PuPd = GPIO_PuPd_NOPULL;
+	GPIO_Init(GPIOA, &GPIO_InitStruct);
+
+	GPIO_PinAFConfig(GPIOA, GPIO_PinSource5, GPIO_AF_SPI1);
+	GPIO_PinAFConfig(GPIOA, GPIO_PinSource6, GPIO_AF_SPI1);
+	GPIO_PinAFConfig(GPIOA, GPIO_PinSource7, GPIO_AF_SPI1);
+
+	GPIO_InitStruct.GPIO_Pin = GPIO_Pin_8;
+	GPIO_InitStruct.GPIO_Mode = GPIO_Mode_OUT;
+	GPIO_InitStruct.GPIO_OType = GPIO_OType_PP;
+	GPIO_InitStruct.GPIO_Speed = GPIO_Speed_40MHz;
+	GPIO_InitStruct.GPIO_PuPd = GPIO_PuPd_UP;
+	GPIO_Init(GPIOA, &GPIO_InitStruct);
+
+	RCC_APB2PeriphClockCmd(RCC_APB2Periph_SPI1, ENABLE);
+
+	SPI_InitStruct.SPI_Direction = SPI_Direction_2Lines_FullDuplex;
+	SPI_InitStruct.SPI_Mode = SPI_Mode_Master;
+	SPI_InitStruct.SPI_DataSize = SPI_DataSize_8b;
+
+	SPI_InitStruct.SPI_CPOL = SPI_CPOL_High;
+	SPI_InitStruct.SPI_CPHA = SPI_CPHA_2Edge;
+
+	SPI_InitStruct.SPI_NSS = SPI_NSS_Soft;
+	SPI_InitStruct.SPI_BaudRatePrescaler = SPI_BaudRatePrescaler_32;
+	SPI_InitStruct.SPI_FirstBit = SPI_FirstBit_MSB;
+	SPI_Init(SPI1, &SPI_InitStruct);
+
+	SPI_Cmd(SPI1, ENABLE);
+}
+
+void write_reg(uint8_t WriteAddr, uint8_t WriteData )
+{
+	chip_select();
+
+	while(!SPI_I2S_GetFlagStatus(SPI1, SPI_I2S_FLAG_TXE));
+	SPI_I2S_SendData(SPI1, WriteAddr);
+	while(!SPI_I2S_GetFlagStatus(SPI1, SPI_I2S_FLAG_RXNE));
+	SPI_I2S_ReceiveData(SPI1);
+
+	while(!SPI_I2S_GetFlagStatus(SPI1, SPI_I2S_FLAG_TXE));
+	SPI_I2S_SendData(SPI1, WriteData);
+	while(!SPI_I2S_GetFlagStatus(SPI1, SPI_I2S_FLAG_RXNE));
+	SPI_I2S_ReceiveData(SPI1);
+	chip_deselect();
+}
+
+uint8_t read_reg(uint8_t ReadAddr)
+{
+	chip_select();
+	while(!SPI_I2S_GetFlagStatus(SPI1, SPI_I2S_FLAG_TXE));
+	SPI_I2S_SendData(SPI1, ReadAddr);
+	while(!SPI_I2S_GetFlagStatus(SPI1, SPI_I2S_FLAG_RXNE));
+	SPI_I2S_ReceiveData(SPI1); //Clear RXNE bit
+
+	while(!SPI_I2S_GetFlagStatus(SPI1, SPI_I2S_FLAG_TXE));
+	SPI_I2S_SendData(SPI1, 0x00); //Dummy byte to generate clock
+	while(!SPI_I2S_GetFlagStatus(SPI1, SPI_I2S_FLAG_RXNE));
+	uint8_t data =  SPI_I2S_ReceiveData(SPI1);
+
+	chip_deselect();
+	return data;
+}
+
+void read_regs( uint8_t ReadAddr, uint8_t *ReadBuf, unsigned int Bytes )
+{
+    uint16_t  i = 0;
+    for(i=0; i<Bytes; i++){
+    	ReadBuf[i] = read_reg(ReadAddr+i);
+    }
+}
+
+void chip_select(void){
+	GPIO_ResetBits(GPIOA, GPIO_Pin_8);
+}
+void chip_deselect(void){
+	GPIO_SetBits(GPIOA, GPIO_Pin_8);
+}
+
+
