+++ src/stm32l1xx_it.c
@@ -135,10 +135,7 @@ void PendSV_Handler(void)
   */
 void SysTick_Handler(void)
 {
-	/*  TimingDelay_Decrement(); */
-#ifdef USE_STM32L_DISCOVERY
-  TimingDelay_Decrement();
-#endif
+	delay_decrement();
 }
 
 /******************************************************************************/
